+++ aiverify-portal/app/canvas/components/drawers/reportInputBlocks.tsx
@@ -1,6 +1,6 @@
 import { RiFileTextLine, RiFileTextFill } from '@remixicon/react';
 import { InputBlock } from '@/app/types';
-import { Button } from '@/lib/components/Button';
+import { Button } from '@/lib/components/TremurButton';
 import {
   DrawerClose,
   DrawerContent,
